fix(runtime): refactored runtime to sane directory layout.
