+++ runtime/src/app/components/header/header.component.scss
@@ -1,4 +1,4 @@
-@import '../../assets/stylesheets/base';
+@import '../../../assets/stylesheets/base';
 
 $pf-cust-navbar-text-size:$font-size-large;
