+++ runtime/src/app/components/login/login.component.scss
@@ -1,4 +1,4 @@
-@import '../../assets/stylesheets/base';
+@import '../../../assets/stylesheets/base';
 
 .login-pf{
   background-color:$color-pf-black;
